[MOD]修改作业需要的node从int变为float，表明单个作业条目需要的节点数量

+++ job/job_type.py
@@ -10,7 +10,7 @@ from pydantic import BaseModel
 
 class SingleItemAllocation(BaseModel):
     '单个待处理条目所需要的资源'
-    node: int
+    node: float
     memory: int
     unit: str
